refactor: export const memos; move conflict helpers to SectionTableBody/helpers
